+++ conftest.py
@@ -1,0 +1,4 @@
+"""Root conftest.py"""
+
+# Don't collect e2e tests; only run when specified over CLI.
+collect_ignore = ["tests/e2e_tests"]