+++ src/gh_profiler/cli.py
@@ -19,7 +19,8 @@ from .utils.profile_data import profile_data as pdata
     help="Generate a workflow that will automatically run `gh-profiler --concise` for every new PR and issue.",
 )
 @click.option("--redact", is_flag=True, help="Redact identifying information.")
-def main(target, concise, generate_workflow, redact):
+@click.option("--benchmark-fetch", is_flag=True, help="Benchmark the code that fetches external data.")
+def main(target, concise, generate_workflow, redact, benchmark_fetch):
     """Examine a GitHub user's profile, to help quickly decide how much to invest in their contributions.
 
     You can target a GitHub username, or a PR/issue number from the repository you're working in.
@@ -35,11 +36,12 @@ def main(target, concise, generate_workflow, redact):
     $ python -m gh_profiler ehmatthes
       ...
     """
-    _validate_command(target, concise, generate_workflow, redact)
+    _validate_command(target, concise, generate_workflow, redact, benchmark_fetch)
 
     # Parse CLI options.
     pdata.concise = concise
     pdata.redact = redact
+    pdata.benchmark_fetch = benchmark_fetch
     pdata.generate_workflow = generate_workflow
 
     # If --generate-workflow was passed, go straight to that work.
@@ -60,7 +62,7 @@ def main(target, concise, generate_workflow, redact):
         cli_utils.process_pr_issue_num(pr_issue_num)
         gh_profiler.main()
 
-def _validate_command(target, concise, generate_workflow, redact):
+def _validate_command(target, concise, generate_workflow, redact, benchmark_fetch):
     """Validate arguments that were passed in the CLI call."""
 
     # If target is not passed, --generate-workflow must be passed.