+++ src/gh_profiler/utils/profile_utils.py
@@ -2,10 +2,12 @@
 
 import json
 import sys
+from concurrent.futures import ThreadPoolExecutor
 from datetime import datetime as dt
 from datetime import timedelta
 from datetime import timezone as tz
 from textwrap import dedent
+from time import perf_counter
 
 from . import infra_utils
 from .profile_data import profile_data as pdata
@@ -27,15 +29,39 @@ def ensure_gh():
 
 
 def get_data():
-    """Get all data we'll need from GitHub."""
+    """Get all data we'll need from GitHub.
+    
+    Fetch all data we'll need, then parse it into the data structures that
+    can be analyzed and processed.
+    """
+    # This is the first call, and it checks if the user exists. This call needs
+    # to happen before all others.
     _get_profile_dict()
-    _get_socials()
-    _get_pr_activity()
-    _get_issue_activity()
+
+    # Fetch data. This can all be done in parallel. The benchmarking is here
+    # because this is the slowest part of the program, and it's helpful at
+    # times to benchmark just this block of code.
+    ts_before = perf_counter()
+    with ThreadPoolExecutor() as executor:
+        socials_future = executor.submit(_fetch_socials)
+        pr_activity_future = executor.submit(_fetch_pr_activity)
+        issue_activity_future = executor.submit(_fetch_issue_activity)
+
+        socials_str = socials_future.result()
+        pr_activity_str = pr_activity_future.result()
+        issue_activity_str = issue_activity_future.result()
+
+    ts_after = perf_counter()
+    if pdata.benchmark_fetch:
+        print(f"Fetch data: {ts_after - ts_before:.2f} seconds")
+
+    # Parse data. This should only happen after all data has been fetched.
+    _parse_socials(socials_str)
+    _parse_pr_activity(pr_activity_str)
+    _parse_issue_activity(issue_activity_str)
 
 
 # --- Helper functions ---
-
 
 def _get_profile_dict():
     """Get all the profile information we'll need."""
@@ -59,14 +85,17 @@ def _get_profile_dict():
     if pdata.profile_dict["created_at"] is None:
         sys.exit(f"GitHub user '{pdata.username}' not found.")
 
-def _get_socials():
-    """Get social media accounts from user's profile.
+def _fetch_socials():
+    """Fetch social media accounts from user's profile.
     
     Social media accounts from profiles are a separate endpoint, so I believe
     they require an additional API call.
     """
     cmd = f"gh api users/{pdata.username}/social_accounts"
-    socials_str = infra_utils.run_cmd(cmd)
+    return infra_utils.run_cmd(cmd)
+
+def _parse_socials(socials_str):
+    """Parse the data string returned from _fetch_socials()."""
     try:
         pdata.socials = json.loads(socials_str)
     except json.decoder.JSONDecodeError:
@@ -75,8 +104,8 @@ def _get_socials():
         sys.exit(msg)
 
 
-def _get_pr_activity():
-    """Get information about recent PR activity."""
+def _fetch_pr_activity():
+    """Fetch information about recent PR activity."""
     cutoff = (dt.now(tz.utc) - timedelta(days=21)).date().isoformat()
 
     pr_query = _get_pr_query()
@@ -84,9 +113,12 @@ def _get_pr_activity():
         f"author:{pdata.username} is:pull-request is:public created:>={cutoff}"
     )
     cmd = f"gh api graphql -f query='{pr_query}' -F q='{search_query}' -F n=100"
+    return infra_utils.run_cmd(cmd)
 
+def _parse_pr_activity(pr_activity_str):
+    """Parse the data returned by _fetch_pr_activity()."""
     try:
-        data = json.loads(infra_utils.run_cmd(cmd))
+        data = json.loads(pr_activity_str)
     except json.decoder.JSONDecodeError:
         msg = "Couldn't get recent PR activity. The gh CLI may have timed out."
         msg += "\n  You may want to try running the command again."
@@ -102,19 +134,16 @@ def _get_pr_activity():
     )
 
 
-def _get_issue_activity():
-    """Get target user's recent public issue activity."""
+def _fetch_issue_activity():
+    """Fetch target user's recent public issue activity."""
     cutoff = (dt.now(tz.utc) - timedelta(days=21)).date().isoformat()
     gh_call = _get_gh_issues_call(pdata.username, cutoff)
-    try:
-        issue_activity = infra_utils.run_cmd(gh_call)
-    except ValueError:
-        msg = "Couldn't get recent issue activity. The gh CLI may have timed out."
-        msg += "\n  You may want to try running the command again."
-        sys.exit(msg)
+    return infra_utils.run_cmd(gh_call)
 
+def _parse_issue_activity(issue_activity_str):
+    """Parse data returned by _fetch_issue_activity()."""
     try:
-        pdata.issue_activity = json.loads(issue_activity)["data"]["search"]
+        pdata.issue_activity = json.loads(issue_activity_str)["data"]["search"]
     except (json.decoder.JSONDecodeError, KeyError):
         msg = "Couldn't get recent issue activity. The gh CLI may have timed out."
         msg += "\n  You may want to try running the command again."